+++ connector/connect/client/jvm/src/test/scala/org/apache/spark/sql/test/ConnectFunSuite.scala
@@ -16,12 +16,13 @@
  */
 package org.apache.spark.sql.test
 
+import java.nio.file.Path
+
+import org.scalatest.funsuite.AnyFunSuite // scalastyle:ignore funsuite
+
 import org.apache.spark.connect.proto
 import org.apache.spark.sql.Column
 import org.apache.spark.sql.connect.ColumnNodeToProtoConverter
-
-import java.nio.file.Path
-import org.scalatest.funsuite.AnyFunSuite // scalastyle:ignore funsuite
 
 /**
  * The basic testsuite the client tests should extend from.
